ALL now routes only from known midi devices
This allows us to omit devices by commenting them out from the config. There's
likely a need to have two separate concepts - one for all devices (regardless
of whether they are in config), and one for all devices that are in the config.
For our use case, this allows us to avoid sending data to loopback interfaces
that can cause messages to bounce around endlessly.