Detect midi device changes
This adds asyncio task support to allow us to periodically check the list of midi devices to determine if there's a change. If so, a MidiDeviceChangeException is raised, causing the async runner to cancel and for the app to re-initialize all ports, etc before re-entering the asyncio task logic. Unfortunately there are no midi device change callbacks we can simply subscribe to.

This does make the app much more complicated as we not only have midi inputs coming in on multiple threads, but we also have asyncio cooperating parallelism. But it should make the logic performant while still allowing for periodic midi device change detection.

I also cleaned up the port closing logic in an exception by iteratively tracking ports. that way if an exception is raised in the middle of opening ports, all previously opened ports will be properly closed.