Add support for using short names without port numbers.
This involves parsing out the port number when reading port names from
mido. When opening ports, mido supports short or long names. When coming
up with a list of available ports, we add both short and long names so
that our checks for available names successfully return valid short
names.

While this allows us to identify devices by short name when the port
numbers are not stable, it does not maintain the ability to distinguish
multiple devices with the same short name. Further work will be needed
prior to merging.