Refactor identifier to port assignment
We now perform a prioritized greedy algorithm to assign available midi
ports to configs. This allows multiple configs to use the same device
name (when multiple of the same device are connected) and guarantees
that each one will receive a different device. Since OS-assigned port
numbers are not stable, the decision of assignment is more or less
arbitrary. This is a known limitation. Powering devices in a specific
sequence _may_ make port assignments more reliable. Otherwise,
additional work will be needed to allow dynamic reassignment.

The new algorithm also allows us to simplify a fair amount of logic
around mapping identifiers and port infos, determining used port names,
etc.

The one downside is that it now opens all ports specified in config
proactively. Previously it would only open a port if it was specified
_and_ was used in a mapper. I think both the likelihook of specifying
unused ports as well as the minimal cost of having extra unused ports
makes this a tradeoff I'm comfortable with.

I also made transforms their own class with names so that we can more
easily print debug information about them.